cloud-run/security-dashboard/requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.11.4
gunicorn==21.2.0